+++ cmd/wordrow/wordmap.go
@@ -58,41 +58,24 @@ func processMapFile(
 	return wordmap.AddFile(&content, format)
 }
 
-// Open the map file specified by `fileArgument` and add its mapping to the
-// `wordmap`. If the file cannot be opened or processing failed the function
-// returns an error.
-func openAndProcessMapFile(
-	fileArgument string,
-	wordmap *wordmaps.WordMap,
-) error {
-	filePath, format := parseMapFileArgument(fileArgument)
+// Opens the file provided by the handler and add its mapping to the `wordmap`.
+// If the file cannot be opened or processing failed the function returns an
+// error.
+func openAndProcessMapFileWith(wordmap *wordmaps.WordMap) fileHandler {
+	return func(fileArgument string) error {
+		filePath, format := parseMapFileArgument(fileArgument)
 
-	logger.Debugf("Opening '%s' as a '%s' formatted map file", filePath, format)
-	handle, err := fs.OpenFile(filePath, fs.OReadOnly)
-	if err != nil {
-		return err
-	}
-
-	defer handle.Close()
-	return processMapFile(handle, format, wordmap)
-}
-
-// Open the map files specified by `filePaths` and add their mapping to the
-// `wordmap`. If any map file is invalid this function will return an error
-// immediately (with a partially updated `wordmap`).
-func openAndProcessMapFiles(
-	filePaths []string,
-	wordmap *wordmaps.WordMap,
-) error {
-	for _, filePath := range filePaths {
-		logger.Debugf("Processing '%s' as a map file", filePath)
-		err := openAndProcessMapFile(filePath, wordmap)
+		logger.Debugf("Opening '%s' as a '%s' formatted map file", filePath, format)
+		handle, err := fs.OpenFile(filePath, fs.OReadOnly)
 		if err != nil {
 			return err
 		}
-	}
 
-	return nil
+		defer handle.Close()
+
+		logger.Debugf("Processing '%s' as a map file", filePath)
+		return processMapFile(handle, format, wordmap)
+	}
 }
 
 // Add a CLI defined mapping to the `wordmap`. If the mapping is invalid this
@@ -101,32 +84,32 @@ func processInlineMapping(mapping string, wordmap *wordmaps.WordMap) error {
 	return wordmap.AddFile(&mapping, "csv")
 }
 
-// Add all CLI defined mappings to the `wordmap`. If any mapping is invalid this
-// function will return an error immediately (with a partially updated
-// `wordmap`).
-func processInlineMappings(mappings []string, wordmap *wordmaps.WordMap) error {
+// Add all CLI defined mappings to the `wordmap`. Any error that occurs is
+// returned after all mappings have been processed.
+func processInlineMappings(
+	mappings []string,
+	wordmap *wordmaps.WordMap,
+) (errs []error) {
 	for _, mapping := range mappings {
 		logger.Debugf("Processing '%s' as a CLI specified mapping", mapping)
 		err := processInlineMapping(mapping, wordmap)
 		if err != nil {
-			return err
+			errs = append(errs, err)
 		}
 	}
 
-	return nil
+	return errs
 }
 
-// Get a WordMap for the specified `mapFiles` and `inlineMappings`. If either
-// contains an invalid mapping this function will return an error immediately.
+// Get a WordMap for the specified `mapFiles` and `inlineMappings`. Any error
+// that occurs is returned after both have been processed. In case of any error
+// the `wordmap` that is returned represents only the arguments that could be
+// successfully processed.
 func getWordMap(
 	mapFiles []string,
 	inlineMappings []string,
-) (wordmap wordmaps.WordMap, err error) {
-	err = openAndProcessMapFiles(mapFiles, &wordmap)
-	if err != nil {
-		return wordmap, err
-	}
-
-	err = processInlineMappings(inlineMappings, &wordmap)
-	return wordmap, err
+) (wordmap wordmaps.WordMap, errs []error) {
+	errs = forEach(mapFiles, openAndProcessMapFileWith(&wordmap))
+	errs = append(errs, processInlineMappings(inlineMappings, &wordmap)...)
+	return wordmap, errs
 }