+++ internal/fs/paths.go
@@ -16,12 +16,9 @@ func GetExt(path string) string {
 	return filepath.Ext(path)
 }
 
-// ResolveGlobs resolves any number of globs or file paths into distinct file paths.
-//
-// The function sets the error if at least one malformed pattern is found. Only
-// the last malformed pattern is reported. The list of paths will contain all
-// paths for valid not-malformed patterns.
-func ResolveGlobs(patterns ...string) (paths []string, rerr error) {
+// ResolveGlobs resolves any number of globs or file paths into distinct file
+// paths. The function returns an error for every invalid pattern.
+func ResolveGlobs(patterns ...string) (paths []string, errs []error) {
 	for _, pattern := range patterns {
 		if !globExpr.MatchString(pattern) {
 			paths = append(paths, pattern)
@@ -30,11 +27,11 @@ func ResolveGlobs(patterns ...string) (paths []string, rerr error) {
 
 		matches, err := filepathx.Glob(pattern)
 		if err != nil {
-			rerr = errors.Newf("Malformed pattern (%s)", pattern)
+			errs = append(errs, errors.Newf("Malformed pattern (%s)", pattern))
 		} else {
 			paths = append(paths, matches...)
 		}
 	}
 
-	return paths, rerr
+	return paths, errs
 }