+++ CHANGELOG.md
@@ -7,10 +7,19 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [0.6.0-beta] - 2020-08-18
+
+### Breaking Changes
+
+- Continue on non-blocking errors. ([#112])
+
 ### Features
 
 - Add support for CLI arguments with "=". ([#102])
 - Add support for many-to-one mapping syntax on the CLI. ([#118])
+- Add `--strict`/`-S` flag to stop for any error. ([#112])
 
 ### Bug Fixes
 
@@ -134,6 +143,7 @@ Versioning].
 [#107]: https://github.com/ericcornelissen/wordrow/pull/107
 [#108]: https://github.com/ericcornelissen/wordrow/pull/108
 [#110]: https://github.com/ericcornelissen/wordrow/pull/110
+[#112]: https://github.com/ericcornelissen/wordrow/pull/112
 [#114]: https://github.com/ericcornelissen/wordrow/pull/114
 [#115]: https://github.com/ericcornelissen/wordrow/pull/115
 [#116]: https://github.com/ericcornelissen/wordrow/pull/116